+++ src/roman_disperser/__init__.py
@@ -9,6 +9,8 @@ from . import demo_utils
 from . import psf_utils
 from . import psf_model
 from . import catalog
+from . import hydrate
+from . import paths
 from . import refdata
 from . import sersic
 
@@ -22,6 +24,8 @@ __all__ = [
     "psf_utils",
     "psf_model",
     "catalog",
+    "hydrate",
+    "paths",
     "refdata",
     "sersic",
 ]